chore: refactor backend creation
